Split editors into a separate library
