+++ lib/widgets.dart
@@ -15,6 +15,7 @@
 /// This library may depend on the data, services, and models library.   
 library widgets;
 
+export "src/widgets/atomic/editors.dart";
 export "src/widgets/atomic/video_feed.dart";
 
 export "src/widgets/generic/feeds.dart";
